requirements.txt
ai.engine==0.0.1
aiohttp==3.9.5
aiosignal==1.3.1
apispec==6.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==23.2.0
bech32==1.2.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cosmpy==0.9.2
ecdsa==0.18.0
Flask==3.0.3
Flask-Cors==3.0.8
Flask-JWT-Extended==4.6.0
frozenlist==1.4.1
googleapis-common-protos==1.63.0
grpcio==1.62.2
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
marqo==3.2.1
minio==7.2.5
msgpack==1.0.8
multidict==6.0.5
packaging==24.0
protobuf==4.25.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodome==3.20.0
pydantic==1.10.15
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
six==1.16.0
typing_extensions==4.11.0
uagents==0.11.1
urllib3==1.26.18
uvicorn==0.20.0
websockets==10.4
Werkzeug==3.0.2
yarl==1.9.4
